==> Training the Model/01 Download all data .ipynb <==
!curl "https://singlecell.broadinstitute.org/single_cell/api/v1/search/bulk_download?accessions=SCP542&auth_code=QF85ofar&directory=all"  -o cfg.txt; curl -K cfg.txt && rm cfg.txt

from zipfile import ZipFile
import os

os.listdir('SCP542/other')

# CCLE_scRNAseq_github.zip represent the data of the mapped cell lines with cells in a 3-d Matrix

ccle_zip = 'SCP542/other/CCLE_scRNAseq_github.zip'

#extracting the zip file

with ZipFile(ccle_zip, 'r') as zip:
    # printing all the contents of the zip file
    zip.printdir()
  
    # extracting all the files
    print('Extracting all the files now...')
    zip.extractall()
    print('Done!')
