==> Training the Model/04 Pre-Processing Min_max Normalization.ipynb <==
import tensorflow as tf
import numpy as np
import pandas as pd
import os 
import matplotlib.pyplot as plt
from tensorflow.keras.utils import to_categorical

os.listdir('TTV')

%%time
TRAIN = pd.read_csv('TTV/TRAIN.csv')

%%time
TEST = pd.read_csv('TTV/TEST.csv')

%%time
VALIDATE = pd.read_csv('TTV/VALIDATE.csv')

TRAIN

TEST

TRAIN = TRAIN.sample(frac=1);
VALIDATE = VALIDATE.sample(frac=1);
TEST = TEST.sample(frac=1);

TRAIN

train_labels_1 = TRAIN.pop('Cell_Line')
train_labels_2 = TRAIN.pop('Cancer_Type')
test_labels_1 = TEST.pop('Cell_Line')
test_labels_2 = TEST.pop('Cancer_Type');
validate_labels_1 = VALIDATE.pop('Cell_Line');
validate_labels_2 = VALIDATE.pop('Cancer_Type');

columns = np.array(VALIDATE.columns)

np.save('columns.npy',columns)

lab1_dict = {};
i = 0 ;
for each in train_labels_1.unique():
    lab1_dict[each] = i
    i += 1
    

lab2_dict = {};
i = 0 ;
for each in train_labels_2.unique():
    lab2_dict[each] = i
    i += 1

lab2_dict

lab1_arr = []
for key,values in lab1_dict.items():
    lab1_arr.append([key,values])

lab1_arr = np.array(lab1_arr)
np.save('lab1_arr', lab1_arr)

lab2_arr = []
for key,values in lab2_dict.items():
    lab2_arr.append([key,values])
    
lab2_arr = np.array(lab2_arr);
np.save('lab2_arr', lab2_arr);

lab1_arr = np.load('lab1_arr.npy')
lab2_arr = np.load('lab2_arr.npy')

for key,value in lab1_arr:
    train_labels_1 = train_labels_1.replace({key:value})
    test_labels_1 = test_labels_1.replace({key:value})
    validate_labels_1 = validate_labels_1.replace({key:value})

for key,value in lab2_arr:
    train_labels_2 = train_labels_2.replace({key:value})
    test_labels_2 = test_labels_2.replace({key:value})
    validate_labels_2 = validate_labels_2.replace({key:value})

len(validate_labels_2.unique())

test_labels_2.unique()

TRAIN.isnull().any()

TRAIN = TRAIN.fillna(0)

VALIDATE = VALIDATE.fillna(0);

TEST = TEST.fillna(0);

TEST.isnull().any()

VALIDATE.isnull().any()

mean_std_col = np.load('mean_std_cols.npy',allow_pickle=True)

df_train_z = TRAIN.copy()
df_test_z = TEST.copy()
df_val_z = VALIDATE.copy()
for key,value in mean_std_col:
    Mean,Std = value[0],value[1]
    df_train_z[key] = (df_train_z[key] - Mean) / Std 
    df_test_z[key] = (df_test_z[key] - Mean) / Std ;
    df_val_z[key] = (df_val_z[key] - Mean) / Std 
    
     

l = df_test_z['RP11-34P13.7']

train_features_z = pd.DataFrame(df_train_z).to_numpy()
test_features_z = pd.DataFrame(df_test_z).to_numpy()
validate_features_z = pd.DataFrame(df_val_z).to_numpy()

np.save('train_features_z',train_features_z)
np.save('test_features_z',test_features_z)
np.save('validate_features_z',validate_features_z)

test_features_z.shape

min_max_col = np.load('min_max_cols.npy',allow_pickle=True)

df_train_s = TRAIN.copy()
df_test_s = TEST.copy()
df_val_s = VALIDATE.copy()
for key,value in min_max_col:
    Min,Max = value[0],value[1]
    df_train_s[key] = (df_train_s[key] - Min) / (Max - Min)  
    df_test_s[key] = (df_test_s[key] - Min) / (Max - Min) ;
    df_val_s[key] = (df_val_s[key] - Min) / (Max - Min) 

l = df_test_s['RP11-34P13.7']

l.isnull().any()

#for i in l:
    #if i!=0:
        #print(i)

train_features_s = pd.DataFrame(df_train_s).to_numpy()
test_features_s = pd.DataFrame(df_test_s).to_numpy()
validate_features_s = pd.DataFrame(df_val_s).to_numpy()

np.save('train_features_s',train_features_s)
np.save('test_features_s',test_features_s)
np.save('validate_features_s',validate_features_s)

test_features_s.shape

train_l1 = to_categorical(train_labels_1,198) #this is the code to one hot encode the labels 
train_l2 = to_categorical(train_labels_2,22) # 198 and 22 represents how many classes are you going to use this function to one-hot encode
test_l1 = to_categorical(test_labels_1,198)
test_l2 = to_categorical(test_labels_2,22)
validate_l1 = to_categorical(validate_labels_1,198)
validate_l2 = to_categorical(validate_labels_2,22)

train_l1.shape

np.save('train_l1',train_l1 ) #saving the one hot encoded labels 
np.save('train_l2',train_l2 )
np.save('test_l1',test_l1)
np.save('test_l2',test_l2)
np.save('validate_l1',validate_l1)
np.save('validate_l2',validate_l2)

validate_l2.shape

# saving a test set data for example 

Example = TEST.sample(frac=0.01);

Example_lab1 = Example.pop('Cell_Line')

Example_lab2 = Example.pop('Cancer_Type')

Example.to_csv('Example.csv',sep=',', index=False)

Example

Example_lab1 = to_categorical(Example_lab1,198) #this is the code to one hot encode the labels 
Example_lab1 = to_categorical(Example_lab2,22)

np.save('test_l1',test_l1)
np.save('test_l2',test_l2)
