==> Training the Model/06 Model for Cancer Type Prediction.ipynb <==
import tensorflow as tf
import numpy as np
import pandas as pd
import os 
import matplotlib.pyplot as plt
from tensorflow.keras.utils import to_categorical

lab1_arr = np.load('lab1_arr.npy')
lab2_arr = np.load('lab2_arr.npy')

train_features = np.load('train_features_s.npy'); # this should load all the data sets
test_features = np.load('test_features_s.npy');
validate_features = np.load('validate_features_s.npy');
train_l1= np.load('train_l1.npy'); # l1 l2 such are one hot encoded
train_l2= np.load('train_l2.npy');
validate_l1= np.load('validate_l1.npy');
validate_l2= np.load('validate_l2.npy');
test_l2= np.load('test_l2.npy');
test_l1= np.load('test_l1.npy'); 

train_features = np.nan_to_num(train_features)
test_features = np.nan_to_num(test_features)
validate_features = np.nan_to_num(validate_features)

def model2():
    model = tf.keras.Sequential([tf.keras.layers.Dense(32,activation='relu',input_shape=(31015,),name='d1_32'),
                                tf.keras.layers.Dense(64,activation='relu',name='d2_64',kernel_regularizer = 'l2'),
                                tf.keras.layers.Dropout(0.2,name='Dropout1_0.2'),
                                tf.keras.layers.Dense(64,activation='relu',name='d3_64',kernel_regularizer='l2'),
                                tf.keras.layers.Dropout(0.2,name='Dropout2_0.2'),
                                tf.keras.layers.Dense(198,activation='softmax',name='output')
                                ],name='Model_2')
    
    model.compile(loss='categorical_crossentropy',optimizer='adam',metrics=['accuracy'])

    model.summary()
    
    return model

Cancer_Cell_Line_Predictor = model2()

checkpoint_path='Checkpoints_ClineS/model2/c6/cp-0020.ckpt'

Cancer_Cell_Line_Predictor.load_weights(checkpoint_path)

Cancer_Cell_Line_Predictor

Cancer_Cell_Line_Predictor.add(tf.keras.layers.Dense(22,activation='softmax',name='output_cancer_Types'))

Cancer_Type_Predictor = Cancer_Cell_Line_Predictor

History_Cancer_Type = Cancer_Type_Predictor.fit(train_features,train_l2,epochs=8,batch_size=128,validation_data=(validate_features,validate_l2),verbose=2)

Test_Cancer_Type_predictor = Cancer_Type_Predictor.evaluate(test_features, test_l2)

save_path_cancer_type = 'Cancer_Type_Predictor'

Cancer_Type_Predictor.save('CancerTypePredictor.h5')

Cancer_Cell_Line_Predictor = model2()

checkpoint_path='Checkpoints_ClineS/model2/c6/cp-0020.ckpt'

Cancer_Cell_Line_Predictor.load_weights(checkpoint_path)

Cancer_Cell_Line_Predictor.save('CancerCellLinePredictor.h5')

# lets also save the col names
