==> Training the Model/07 Load and Run the model.ipynb <==
import tensorflow as tf
import numpy as np
import pandas as pd
import os 
import matplotlib.pyplot as plt
from tensorflow.keras.utils import to_categorical

lab1_arr = np.load('lab1_arr.npy')
lab2_arr = np.load('lab2_arr.npy')

test_features = np.load('test_features_s.npy');


test_l2= np.load('test_l2.npy');
test_l1= np.load('test_l1.npy'); 

#test_l2[0]

test_features = np.nan_to_num(test_features)

CancerCellLinePredictor = tf.keras.models.load_model('CancerCellLinePredictor.h5')

r1 = CancerCellLinePredictor.evaluate(test_features,test_l1)

CancerTypePredictor = tf.keras.models.load_model('CancerTypePredictor.h5')

r2 = CancerTypePredictor.evaluate(test_features,test_l2)

p1 = CancerCellLinePredictor.predict(test_features[365:396])

len(test_features[365:396])

p1[0].max()

val = np.where(p1[0]==p1[0].max())

val

p1[0][6]

lab1_arr[val][0][0]

def predict(model,dataset,one_hot_arr):
    l = len(dataset)
    p = model.predict(dataset)
    results = {} ;
    for i in range(l):
        val = np.where(p[i] == p[i].max());
        results[i] = one_hot_arr[val][0][0];
        
    return results    
        

re = predict(CancerCellLinePredictor,test_features[365:396],lab1_arr)

rc = predict(CancerTypePredictor,test_features[365:396],lab2_arr)

rc
